requirements.txt
aiofiles==24.1.0
aiogram==3.21.0
aiogram_dialog==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp_socks==0.10.1
aioredis==2.0.1
aioschedule==0.5.2
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.16.1
altgraph==0.17.4
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.10.4
asgiref==3.7.2
async-property==0.2.2
async-timeout==4.0.3
asyncpg==0.30.0
asyncstdlib==3.12.0
attrs==25.3.0
auto-py-to-exe==2.42.0
beautifulsoup4==4.12.2
billiard==4.2.1
blinker==1.7.0
bottle==0.12.25
bottle-websocket==0.2.9
Brotli==1.1.0
browser-cookie3==0.19.1
cachetools
celery==5.5.3
certifi==2025.6.15
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
cryptg==0.5.0.post0
curl-cffi==0.5.10
discord==2.3.2
discord.py==2.3.2
distro==1.9.0
dotenv==0.9.9
duckduckgo_search==4.2
Eel==0.16.0
emoji==2.8.0
et-xmlfile==1.1.0
eval_type_backport==0.2.2
executor==23.2
fake-useragent==1.4.0
fast-depends==2.4.12
fastapi==0.109.0
fasteners==0.19
faststream==0.5.42
Flask==3.0.0
frozenlist==1.7.0
future==0.18.3
g4f==0.2.0.2
gevent==23.9.1
gevent-websocket==0.10.1
google-auth==2.28.1
google-auth-oauthlib==1.2.0
greenlet==3.2.3
gspread==6.0.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.6
jiter==0.10.0
Js2Py==0.74
kandinsky.py==2.4
kombu==5.5.4
loguru==0.7.2
lxml==4.9.4
lz4==4.3.3
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.2
mistralai==1.9.1
mmh3==4.1.0
multidict==6.6.3
nest-asyncio==1.5.8
numpy==1.26.3
oauthlib==3.2.2
openai==1.93.0
openpyxl==3.1.2
opentele==1.15.1
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.0
pefile==2023.2.7
pillow==10.2.0
pip-review==1.3.0
platformdirs==4.1.0
portalocker==3.2.0
prompt_toolkit==3.0.51
propcache==0.3.2
property-manager==3.0
py-arkose-generator==0.0.0.2
pyaes
pyasn1
pyasn1-modules
pycparser==2.21
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==2.11.7
pydantic_core==2.33.2
PyExecJS==1.5.1
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.11
pyjsparser==2.7.1
PyNaCl==1.5.0
pyparsing==3.1.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
Pyrogram==2.0.106
PySocks==1.7.1
pyTelegramBotAPI==4.13.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-socks==2.7.1
pytz==2023.3.post1
pywin32-ctypes==0.2.2
redis==6.2.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9.1
scheduler==0.8.5
selenium==4.16.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.41
starlette==0.35.1
StrEnum==0.4.15
telebot==0.0.5
telegram-session-converter==0.1.3
Telethon==1.40.0
tenacity==9.1.2
TGConvertor==0.0.8
TgCrypto==1.2.5
tqdm==4.67.1
trio==0.23.2
trio-websocket==0.11.1
typeguard==4.1.5
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.2
undetected-chromedriver==3.5.4
urllib3==2.0.5
uvicorn==0.25.0
verboselogs==1.7
vine==5.1.0
vk-api==11.9.9
waitress==2.1.2
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.1
whichcraft==0.6.1
wsproto==1.2.0
yagpt_py==0.0.5
yandexgptlite==0.1.7
yarl==1.20.1
YooMoney==0.1.0
zope.event==5.0
zope.interface==6.1

